Refactor cli
